move Prompts folder out of Commands
