+++ client/Prompts/ConditionReferencePrompt.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
-import { Conditions } from "../../Rules/Conditions";
+import { Conditions } from "../Rules/Conditions";
 import _ = require("lodash");
 import { PromptProps } from "./PendingPrompts";
-import { SubmitButton } from "../../Components/Button";
+import { SubmitButton } from "../Components/Button";
 
 export function ConditionReferencePrompt(
   conditionName: string
