+++ client/Prompts/EditAliasPrompt.tsx
@@ -1,8 +1,8 @@
 import * as React from "react";
 import { PromptProps } from "./PendingPrompts";
-import { Combatant } from "../../Combatant/Combatant";
-import { probablyUniqueString } from "../../../common/Toolbox";
-import { SubmitButton } from "../../Components/Button";
+import { Combatant } from "../Combatant/Combatant";
+import { probablyUniqueString } from "../../common/Toolbox";
+import { SubmitButton } from "../Components/Button";
 import { Field } from "formik";
 import { StandardPromptLayout } from "./StandardPromptLayout";
