+++ client/Prompts/PlayerViewPrompt.tsx
@@ -1,7 +1,7 @@
 import { Field } from "formik";
 import * as React from "react";
-import { Button, SubmitButton } from "../../Components/Button";
-import { env } from "../../Environment";
+import { Button, SubmitButton } from "../Components/Button";
+import { env } from "../Environment";
 import { PromptProps } from "./PendingPrompts";
 
 const promptClassName = "p-launch-player-view";
